+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Non ambient light didn't got turn-off in last-left callback
+
 ## 2.19.4 (2023-10-22)
 
 * (theimo1221) Retrigger publish